% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gale__2013.R
\docType{data}
\name{BAB__GALE__2013}
\alias{BAB__GALE__2013}
\title{BAB}
\format{
A data frame with 1 row and 70 element concentrations:\cr
MgO, SiO2, FeO, CaO, Na2O, Al2O3, TiO2, K2O, P2O5, MnO, Ba, Be, Ce, Co, Cr, Cs, Cu, Dy, Er, Eu, Ga, Gd, Hf, Ho, La, Li, Lu, Mo, Nb, Nd, Ni, Pb, Pr, Rb, Sc, Sm, Sn, Sr, Ta, Tb, Th, Tl, U, V, W, Y, Yb, Zn, Zr, Sr87_Sr86, Nd143_Nd144, Pb206_Pb204, Pb207_Pb204, Pb208_Pb204, Hf176_Hf177, Sm_Nd, Zr_Hf, Ba_Th, Nb_U, Ce_Pb, Nb_Ta, Th_U, Ba_Rb, Ba_Cs, Rb_Cs, K_U, Y_Ho, Zr_Sm, Hf_Nd, Y_Yb

In the article the concentrations ...
}
\usage{
BAB__GALE__2013
}
\description{
A data set containing the composition of back-arc basin basalts as given by \insertCite{gale2013mean;textual}{georefdatar}
}
\references{
{
\insertRef{gale2013mean}{georefdatar}
}
}
\keyword{datasets}
