% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rudnick_Gao__2014.R
\docType{data}
\name{CC_Lower__Rudnick_Gao__2014}
\alias{CC_Lower__Rudnick_Gao__2014}
\title{Lower Continental Crust}
\format{
A data frame with 1 row and 84 element concentrations. These elements are:\cr
SiO2, TiO2, Al2O3, FeO*, MnO, MgO, CaO, Na2O, K2O, P2O5, Li, Be, B, N, F, S, Cl, Sc, V, Cr, Co, Ni, Cu, Zn, Ga, Ge, As, Se, Br, Rb, Sr, Y, Zr, Nb, Mo, Ru, Pd, Ag, Cd, In, Sn, Sb, I, Cs, Ba, La, Ce, Pr, Nd, Sm, Eu, Gd, Tb, Dy, Ho, Er, Tm, Yb, Lu, Hf, Ta, W, Re, Os, Ir, Pt, Au, Hg, Tl, Pb, Bi, Th, U, Nb/Ta, Zr/Hf, Th/U, K/U, La/Yb, Rb/Cs, K/Rb, La/Ta, Mg#, Eu/Eu*, Heat production
}
\usage{
CC_Lower__Rudnick_Gao__2014
}
\description{
A data set containing the composition of the \emph{Lower Continental Crust} as
recommended by \insertCite{Rudnick2014;textual}{georefdatar}. This article is
a revision of the previous work \insertCite{Rudnick2003;textual}{georefdatar}.
}
\details{
In this work the concentrations of the major elements (as oxides) are given
in wt\%. The concentrations of all other elements are given in ug/g (ppm) or
ng/g (ppb). For the sake of unity the values given in ppb where converted
to ppm using \code{ppm= ppb/1000}.\cr
The listed values for the major elements (oxides) are in wt\% and all other
elements are given in ppm. Heat production is given in mW/m^3.
}
\references{
{
\insertRef{Rudnick2014}{georefdatar}
\insertRef{Rudnick2003}{georefdatar}
}
}
\keyword{datasets}
