% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM__Salters_Stracke__2004.R
\docType{data}
\name{DM__Salters_Stracke__2004}
\alias{DM__Salters_Stracke__2004}
\title{Depleted mantle}
\format{
A data frame with 1 row and 80 element concentrations in ppm:\cr
H2O, He, Li, Be, B, CO2, N2, F, Ne, Na2O, MgO, Al2O3, SiO2, P, S, Cl, Ar, K, CaO, Sc, Ti, V, Cr, Mn, FeO, Co, Ni, Cu, Zn, Ga, Ge, As, Se, Kr, Rb, Sr, Y, Zr, Nb, Mo, Ru, Rh, Pd, Ag, Cd, In, Sn, Sb, Te, Xe, Cs, Ba, La, Ce, Pr, Nd, Sm, Eu, Gd, Tb, Dy, Ho, Er, Tm, Yb, Lu, Hf, Ta, W, Re, Os, Ir, Pt, Au, Hg, Tl, Pb, Bi, Th, U
}
\usage{
DM__Salters_Stracke__2004
}
\description{
A data set containing the estimated composition of the Depleted Mantle as
given by \insertCite{Salters2004;textual}{georefdatar}.
}
\details{
Please refer to the paper for the units.
}
\references{
{
\insertRef{Salters2004}{georefdatar}
}
}
\keyword{datasets}
