% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sun_McDonough__1989.R
\docType{data}
\name{EMORB__Sun_McDounough__1989}
\alias{EMORB__Sun_McDounough__1989}
\title{E-type MORB}
\format{
A data frame with 1 row and 36 element concentrations in ppm:\cr
Cs, Tl, Rb, Ba, W, Th, U, Nb, Ta, K, La, Ce, Pb, Pr, Mo, Sr, P, Nd, F, Sm, Zr, Hf, Eu, Sn, Sb, Ti, Gd, Tb, Dy, Li, Y, Ho, Er, Tm, Yb, Lu
}
\usage{
EMORB__Sun_McDounough__1989
}
\description{
A data set containing the element concentrations in the E-type MORB as
given by \insertCite{sun1989chemical;textual}{georefdatar}.
}
\references{
{
\insertRef{sun1989chemical}{georefdatar}
}
}
\keyword{datasets}
