% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcrefer-package.R
\docType{package}
\name{gcrefer-package}
\alias{gcrefer}
\alias{gcrefer-package}
\title{gcrefer: Provides geochmical reference datasets}
\description{
The packe provides some geochmical refernce datasets. These datasets are commonly used for geochemical calculations and plot. The package provides reference dataset libe the compositions of some reservoirs like the prmitive mantle or chondrites.
}
\references{
\insertAllCited{}
}
\author{
\strong{Maintainer}: Gerald Schuberth-Hlavač \email{abuseki@synapticgap.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{--}
\keyword{Rdpack-manual}
\keyword{from}
\keyword{internal}
\keyword{taken}
