% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefdatar-package.R
\docType{package}
\name{georefdatar-package}
\alias{georefdatar}
\alias{georefdatar-package}
\title{georefdatar: Provides geochemical reference datasets}
\description{
The packe provides some reference data sets used in geoscience. These data sets are commonly used for geochemical calculations and plots. The package provides reference data sets like the compositions of some reservoirs e.g. the primitive mantle or chondrites.
}
\references{
\preformatted{\\insertAllCited\{\}
}
}
\author{
\strong{Maintainer}: Gerald Schuberth-Hlavač \email{abuseki@synapticgap.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{--}
\keyword{Rdpack-manual}
\keyword{from}
\keyword{internal}
\keyword{taken}
