% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefdatar-package.R
\name{georefdatar_package}
\alias{georefdatar_package}
\alias{georefdatar-package}
\title{Geosciences reference data sets in R}
\description{
The package provides some reference data sets that are regularly used in
geosciences. These are a periodic table of the elements, a list of minerals
and some reservoir reference data set (continental crust, mantle, basalts,
...). Some functions for basic queries of the periodic table of the elements
and the list of minerals are also included.\cr
All records have full references, making them citable.
}
\references{
{
\insertNoCite{*}{georefdatar}
\insertAllCited{}
}
}
\author{
Gerald Schuberth-HlavaÄ
}
