% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementSets.R
\docType{data}
\name{Lithophile}
\alias{Lithophile}
\alias{Chalcophile}
\alias{Siderophile}
\alias{Atmophile}
\alias{Goldschmidt}
\title{Goldschmidt's classification of the elements}
\format{
\code{Lithophile}: character vector with 46 elements.

\code{Chalcophile}: character vector with 15 elements.

\code{Siderophile}: character vector with 15 elements.

\code{Atmophile}: character vector with 10 elements.
}
\usage{
Lithophile

Chalcophile

Siderophile

Atmophile
}
\description{
Sets containing the elements classified after Goldschmidt.
}
\details{
The geochemical behavior of the elements is controlled by many factors (e.g.
ionic radius, volatility, redox, ...). Therefore, the elements can be
classified in several ways. A common classification scheme is that developed
by \emph{V. M. Goldschmidt}, which is based on the affinity of elements to form
different types of compounds. The Elements are characterized as:
\enumerate{
\item \strong{lithophile} (rock-loving): elements with a strong affinity for forming
oxides and silicate minerals.
\item \strong{chalcophile}  (copper-loving): elements with a strong affinity for
forming sulfides.
\item \strong{siderophile}  (iron-loving): elements with a strong affinity to form
metals or solid solutions in metals.
\item \strong{atmophile}  (air-loving): elements that exist either uncombined or as
highly volatile compounds.
}
}
\examples{
# List the atmophile elements
Atmophile

# Show the electron configuration of the atmophile elements
pte[pte$Symbol \%in\% Atmophile, c("Symbol", "ElectronConfiguration")]

}
\references{
\insertRef{Marshall1999}{georefdatar}
}
\keyword{datasets}
