% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_Colors.R
\docType{data}
\name{ICS_Colors}
\alias{ICS_Colors}
\title{The CGMW ICS color codes}
\format{
A data frame with 194 rows and the following
11 columns:\cr
\enumerate{
\item \verb{standard sorting order}  ICS' ordering of this entry
\item \verb{Long List (isc:)}        Entries name prefixed by \verb{ics:}
\item \verb{Long List (formatted)}   The (common) name of entry, e.g. 'Holocene'
\item \code{Rank}                    Is the entry a System, Series, Stage, ...
\item \code{Cyan}                    CMYK colors
\item \code{Magenta}
\item \code{Yellow}
\item \code{Black}
\item \code{Red}                     RGB colors
\item \code{Green}
\item \code{Blue}
}
}
\usage{
ICS_Colors
}
\description{
A data set containing the color codes used by the
International Chronostratigraphic Chart by the
International Commission on Stratigraphy (ICS)
\insertCite{Cohen2013}{georefdatar}.
The coloring in this chart follows the
Commission for the Geological Map of the World (CGMW)
\insertCite{Vrielynck2022}{georefdatar}.
}
\references{
{
\insertRef{Cohen2013}{georefdatar}

\insertRef{Vrielynck2022}{georefdatar}
}
}
\keyword{datasets}
