% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_Colors.R
\name{icsColor}
\alias{icsColor}
\title{Get ICS Color for a unit name found in the International Chronostratigraphic
Chart.}
\usage{
icsColor(name, colorModel = "RGB")
}
\arguments{
\item{name}{character. The name of a unit: eontheme to stage}

\item{colorModel}{character. The color model to get the color codes in --
either 'RGB' (default) or 'CMYK'.}
}
\value{
list of the color code in the chosen color model
}
\description{
Retrieve the color code for a given name of an eontheme, eratheme, system,
... from the color codes of the International Chronostratigraphic Chart.
}
\examples{
# Color codes of the Permian in RGB
icsColor("Permian")


}
\seealso{
\link{ICS_Colors} for the full color code table
}
