% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeriodicTable.R
\name{m}
\alias{m}
\title{Get the atomic mass of an element}
\usage{
m(sym)
}
\arguments{
\item{sym}{Symbol of the element as a string}
}
\value{
Atomic mass of element with symbol \code{sym}
}
\description{
Get the atomic mass of an element
}
\examples{
m('H')
m('H')*2+m('O')
}
\seealso{
\link[=pte]{periodic table of elements}
}
