% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minerals.R
\name{minInfo}
\alias{minInfo}
\title{Get information for a minerals}
\usage{
minInfo(sym)
}
\arguments{
\item{sym}{Symbol of the mineral as a string}
}
\value{
Information for a mineral with symbol \code{sym}
}
\description{
Get information for a minerals
}
\examples{
minInfo('Ms')
}
\seealso{
\link[=mins]{List of minerals}
}
