% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minerals.R
\docType{data}
\name{mins}
\alias{mins}
\title{List of Minerals}
\format{
A data frame with 5763 minerals and their names, symbols and chemistry.\cr
Symbol, Name, Chemistry
}
\usage{
mins
}
\description{
IMA approved list of minerals. Retrieved from \emph{RRUFF} \insertCite{Lafuente_2015;textual}{georefdatar}.
}
\references{
{
\insertRef{Lafuente_2015}{georefdatar}

\insertRef{Warr_2021}{georefdatar}
}
}
\keyword{datasets}
