% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sun_McDonough__1989.R
\docType{data}
\name{PM__Sun_McDounough__1989}
\alias{PM__Sun_McDounough__1989}
\title{Primitive mantle}
\format{
A data frame with 1 row and 36 element concentrations in ppm:\cr
Cs, Tl, Rb, Ba, W, Th, U, Nb, Ta, K, La, Ce, Pb, Pr, Mo, Sr, P, Nd, F, Sm, Zr, Hf, Eu, Sn, Sb, Ti, Gd, Tb, Dy, Li, Y, Ho, Er, Tm, Yb, Lu
}
\usage{
PM__Sun_McDounough__1989
}
\description{
A dataset containing the element concentrations in the primitive mantle as
given in \insertCite{sun1989chemical}{georefdatar}.
}
\details{
For lead and cesium the -- in this work -- recommended values for
mantel-normalizing diagrams where taken.
The original values that where given in Tb.1 are (Cs, 0.032) and (Pb, 0.185).
}
\references{
{
\insertRef{sun1989chemical}{georefdatar}
}
}
\keyword{datasets}
