% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeriodicTable.R
\docType{data}
\name{pte}
\alias{pte}
\title{Periodic Table of Elements}
\format{
A data frame with 118 rows and 17 columns.\cr
For each element the following attributes are reported:
AtomicNumber, Symbol, Name, AtomicMass, CPKHexColor, ElectronConfiguration, Electronegativity, AtomicRadius, IonizationEnergy, ElectronAffinity, OxidationStates, StandardState, MeltingPoint, BoilingPoint, Density, GroupBlock, YearDiscovered
}
\usage{
pte
}
\description{
The periodic table of elements as given by PubChem \insertCite{pubChemPTable;textual}{georefdatar}.
}
\references{
{
\insertRef{pubChemPTable}{georefdatar}

\insertRef{Kim_2020}{georefdatar}
}
}
\seealso{
\link{IUPAC_StdAW} for the standard atomic weights of the elements
recommended by \href{https://iupac.org}{IUPAC}
}
\keyword{datasets}
