% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McDonough_Sun__1995.R
\docType{data}
\name{Pyrolite__McDonough_Sun__1995}
\alias{Pyrolite__McDonough_Sun__1995}
\title{Pyrolite}
\format{
A data frame with 1 row and 76 element concentrations in \link{ppm}:\cr
Li, Be, B, C, N, F, Na, Mg, Al, Si, P, S, Cl, K, Ca, Sc, Ti, V, Cr, Mn, Fe, Co, Ni, Cu, Zn, Ga, Ge, As, Se, Br, Rb, Sr, Y, Zr, Nb, Mo, Ru, Rh, Pd, Ag, Cd, In, Sn, Sb, Te, I, Cs, Ba, La, Ce, Pr, Nd, Sm, Eu, Gd, Tb, Dy, Ho, Er, Tm, Yb, Lu, Hf, Ta, W, Re, Os, Ir, Pt, Au, Hg, Tl, Pb, Bi, Th, U

In the article the concentrations of most of the major elements are given in
wt\% and the concentrations of Nb and following are given in ppb.\cr
For the sake of clarity these values where converted to ppm -- so all given
values are in ppm.\cr
This conversion was done using:
\itemize{
\item \verb{ppm= wt\% * 10000}
\item \code{ppm= ppb / 1000}
}
}
\usage{
Pyrolite__McDonough_Sun__1995
}
\description{
A dataset containing the \emph{recommended chemical composition of the of the Silicate Earth-''Pyrolite''} as given in \insertCite{McDonough_1995}{gcrefer}
}
\references{
{
\insertRef{McDonough_1995}{gcrefer}
}
}
\keyword{datasets}
