% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefdatar-package.R
\docType{package}
\name{georefdatar-package}
\alias{georefdatar}
\alias{georefdatar-package}
\title{georefdatar: Provides geochemical reference datasets}
\description{
The package contains some reference datasets that are regularly used in geosciences. These are a periodic table of the elements, a list of minerals and some reservoir reference data set (continental crust, mantle, basalts, ...). Some functions for basic queries of the periodic table of the elements and the list of minerals are also included. All records have full references, making them citable.
}
\references{
\preformatted{\\insertAllCited\{\}
}
}
\author{
\strong{Maintainer}: Gerald Schuberth-Hlavač \email{abuseki@synapticgap.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{--}
\keyword{Rdpack-manual}
\keyword{from}
\keyword{internal}
\keyword{taken}
