% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefdatar-package.R
\name{georefdatar_package}
\alias{georefdatar_package}
\alias{georefdatar-package}
\title{Geosciences Reference Data Sets in R}
\description{
The package includes reference data sets commonly used in
geosciences, such as the standard atomic weights of elements, a periodic
table, a mineral list, reservoir reference datasets (continental crust,
mantle, basalts, etc.), decay constants, and isotopic ratios frequently used
in geochronology. Additionally, the package provides functions for basic
queries of atomic weights and mineral lists.\cr
All datasets have complete references, making them citable.
}
\references{
{ \insertNoCite{*}{georefdatar} \insertAllCited{} }
}
\author{
Gerald Schuberth-HlavaÄ
}
