% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementSets.R
\docType{data}
\name{REE}
\alias{REE}
\alias{Lanthanides}
\alias{LREE}
\alias{HREE}
\title{Rare earth elements -- REE, LREE, HREE, Lanthanides}
\format{
Multiple element sets:
\itemize{
\item Lanthanides   \code{La}--\code{Lu} \insertCite{Connelly2005}{georefdatar}. The geochemical
definition of the rare earth elements -- e.g
\insertCite{Rollinson1993,Marshall1999}{georefdatar}.
\item REE           rare earth metals. Sc, Y and the lanthanides
\insertCite{Connelly2005}{georefdatar}
\item LREE          Light REE
\item HREE          Heavy REE
}

An object of class \code{character} of length 15.

An object of class \code{character} of length 8.

An object of class \code{character} of length 9.
}
\usage{
REE

Lanthanides

LREE

HREE
}
\description{
List of rare earth elements and subsets thereof. The "Red Book"
\insertCite{Connelly2005}{georefdatar} defines the rare earth metals as \code{Sc},
\code{Y} and the lanthanides (\code{La} -- \code{Lu}). In geochemistry, the term "rare earth
elements" is generally limited to the lanthanides-- e.g.
\insertCite{Rollinson1993,Marshall1999}{georefdatar}.\cr
Therefore, it is crucial to consider the context in which this term is used.
}
\section{Functions}{
\itemize{
\item \code{Lanthanides}: Only lanthanides

\item \code{LREE}: Light REE subset of REE

\item \code{HREE}: Heavy REE subset of REE

}}
\examples{
# get information from the periodic table of elements
subset(pte, Symbol \%in\% REE)

}
\references{
\insertRef{Rollinson1993}{georefdatar}

\insertRef{Marshall1999}{georefdatar}

\insertRef{Connelly2005}{georefdatar}
}
\keyword{datasets}
